# docs/swagger.yaml
definitions:
  common_middleware.Response:
    properties:
      data:
        type: object
      errmsg:
        type: string
      errno:
        type: integer
      stack:
        type: object
      trace_id:
        type: object
    type: object
...
  dto.CreateOrUpdateAppReq:
    properties:
      app_id:
        description: 租户id
        type: string
      app_name:
        description: 租户名称
        type: string
      qpd:
        description: 每日请求量限制
        type: integer
      qps:
        description: 每秒请求量限制
        type: integer
      secret:
        description: 租户密钥
        type: string
      white_ips:
        description: ip白名单，支持前缀匹配
        type: string
    type: object
...
  dto.Flow:
    properties:
      today_flow:
        description: 今日流量列表
        items:
          type: integer
        type: array
      yesterday_flow:
        description: 昨日流量列表
        items:
          type: integer
        type: array
    type: object
  dto.ListAppItem:
    properties:
      app_id:
        type: string
      id:
        type: integer
      name:
        type: string
      qpd:
        type: integer
      qps:
        type: integer
      real_qpd:
        type: integer
      real_qps:
        type: integer
      secret:
        type: string
      white_ips:
        type: string
    type: object
  dto.ListAppsRes:
    properties:
      items:
        description: 租户列表
        items:
          $ref: '#/definitions/dto.ListAppItem'
        type: array
      total:
        description: 共计条数
        example: 0
        type: integer
    type: object
  dto.ListServiceItem:
    properties:
      id:
        type: integer
      real_qpd:
        type: integer
      real_qps:
        type: integer
      service_addr:
        type: string
      service_desc:
        type: string
      service_name:
        type: string
      service_type:
        type: integer
      total_node:
        type: integer
    type: object
  dto.ListServicesRes:
    properties:
      items:
        description: 服务列表
        items:
          $ref: '#/definitions/dto.ListServiceItem'
        type: array
      total:
        description: 共计条数
        example: 0
        type: integer
    type: object
...
  po.App:
    properties:
      app_id:
        type: string
      app_name:
        type: string
      create_at:
        type: string
      id:
        type: integer
      is_delete:
        type: integer
      qpd:
        type: integer
      qps:
        type: integer
      secret:
        type: string
      update_at:
        type: string
      white_ips:
        type: string
    type: object
...
info:
  contact: {}
paths:
  /apps:
    get:
      description: 查询租户列表
      operationId: /apps
      parameters:
      - description: keyword
        in: query
        name: keyword
        type: string
      - description: page index
        in: query
        name: page_index
        type: string
      - description: page size
        in: query
        name: page_size
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            allOf:
            - $ref: '#/definitions/common_middleware.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.ListAppsRes'
              type: object
      summary: 查询租户列表接口
      tags:
      - 租户接口
    post:
      consumes:
      - application/json
      description: 创建租户
      operationId: /apps
      parameters:
      - description: create app request body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.CreateOrUpdateAppReq'
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            allOf:
            - $ref: '#/definitions/common_middleware.Response'
            - properties:
                data:
                  $ref: '#/definitions/po.App'
              type: object
      summary: 创建租户接口
      tags:
      - 租户接口
  /apps/{app_id}:
    delete:
      description: 删除租户
      operationId: /apps/{app_id}
      parameters:
      - description: app id
        in: path
        name: app_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            allOf:
            - $ref: '#/definitions/common_middleware.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: 删除租户接口
      tags:
      - 租户接口
    get:
      description: 查询租户详情
      operationId: /apps/{app_id}
      parameters:
      - description: app id
        in: path
        name: app_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            allOf:
            - $ref: '#/definitions/common_middleware.Response'
            - properties:
                data:
                  $ref: '#/definitions/po.App'
              type: object
      summary: 查询租户详情接口
      tags:
      - 租户接口
    put:
      consumes:
      - application/json
      description: 更新租户
      operationId: /apps/{app_id}
      parameters:
      - description: app id
        in: path
        name: app_id
        required: true
        type: string
      - description: update app request body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.CreateOrUpdateAppReq'
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            allOf:
            - $ref: '#/definitions/common_middleware.Response'
            - properties:
                data:
                  $ref: '#/definitions/po.App'
              type: object
      summary: 更新租户接口
      tags:
      - 租户接口
...
  /dashboard/flow/apps/{app_id}:
    get:
      description: 查询租户流量
      operationId: /flow/apps/{app_id}
      parameters:
      - description: app id
        in: path
        name: app_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            allOf:
            - $ref: '#/definitions/common_middleware.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.Flow'
              type: object
      summary: 查询租户流量接口
      tags:
      - 数据接口
  /dashboard/flow/services/{service_id}:
    get:
      description: 查询服务流量
      operationId: /flow/services/{service_id}
      parameters:
      - description: service id
        in: path
        name: service_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            allOf:
            - $ref: '#/definitions/common_middleware.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.Flow'
              type: object
      summary: 查询服务流量接口
      tags:
      - 数据接口
...
  /services:
    get:
      description: 查询服务列表
      operationId: /services
      parameters:
      - description: keyword
        in: query
        name: keyword
        type: string
      - description: page index
        in: query
        name: page_index
        type: string
      - description: page size
        in: query
        name: page_size
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            allOf:
            - $ref: '#/definitions/common_middleware.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.ListServicesRes'
              type: object
      summary: 查询服务列表接口
      tags:
      - 服务接口
...
swagger: "2.0"